voc 상세 조회 작업

+++ src/main/java/com/example/java/was/domain/VocDto.java
@@ -10,6 +10,6 @@ public class VocDto {
 	private String userKey;
 	private String reason;
 	private String target;
-	private Integer workerId;
-	private Integer vendorId;
+	private Long workerId;
+	private Long vendorId;
 }
